# Inline PackedPixel primitive methods