… src/ImageSharp/Colors/PackedPixel/Bgra5551.cs
… src/ImageSharp/Colors/PackedPixel/Byte4.cs
… src/ImageSharp/Colors/PackedPixel/HalfSingle.cs
… src/ImageSharp/Colors/PackedPixel/HalfVector2.cs
… src/ImageSharp/Colors/PackedPixel/HalfVector4.cs
… src/ImageSharp/Colors/PackedPixel/NormalizedByte2.cs
… src/ImageSharp/Colors/PackedPixel/NormalizedByte4.cs
… src/ImageSharp/Colors/PackedPixel/NormalizedShort2.cs
… src/ImageSharp/Colors/PackedPixel/NormalizedShort4.cs
… src/ImageSharp/Colors/PackedPixel/Rg32.cs
… src/ImageSharp/Colors/PackedPixel/Rgba1010102.cs
… src/ImageSharp/Colors/PackedPixel/Rgba64.cs
… src/ImageSharp/Colors/PackedPixel/Short2.cs
… src/ImageSharp/Colors/PackedPixel/Short4.cs