### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/JimBobSquarePants/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practise as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
![InlineAll](https://i.imgflip.com/1j4ajz.jpg)

Specially the ones that do packing/unpacking!

However, I'm not sure if all CLR implementations are capable of inlining methods on generic `TColor` arguments. Definitely works on the one that executes benchmarks on Windows :P

### Resize benchmarks:
**Before**
```
                      Method |       Mean |    StdDev | Scaled | Scaled-StdDev | Allocated |
---------------------------- |----------- |---------- |------- |-------------- |---------- |
     'System.Drawing Resize' | 53.3980 ms | 1.1966 ms |   1.00 |          0.00 |     512 B |
         'ImageSharp Resize' | 35.9191 ms | 0.2008 ms |   0.67 |          0.02 | 186.06 kB |
 'ImageSharp Compand Resize' | 59.7976 ms | 0.2730 ms |   1.12 |          0.02 | 186.74 kB |
```
**After**
```
                      Method |       Mean |    StdDev | Scaled | Scaled-StdDev | Allocated |
---------------------------- |----------- |---------- |------- |-------------- |---------- |
     'System.Drawing Resize' | 52.0138 ms | 0.5627 ms |   1.00 |          0.00 |     512 B |
         'ImageSharp Resize' | 28.5017 ms | 0.3330 ms |   0.55 |          0.01 |  186.2 kB |
 'ImageSharp Compand Resize' | 57.4592 ms | 0.9292 ms |   1.10 |          0.02 | 186.57 kB |
```