@antonfirsov 2017-02-07
Thanks! Almost did the same merge :)