@JimBobSquarePants 2017-02-07
I'm happy to have this merged. Actually planned to do the same after seeing your comment in the benchmarking code.